backend/pyproject.toml
[project]
name = "storyhub-backend"
version = "0.1.0"
description = "StoryHub local backend service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "aiosqlite>=0.21.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt>=4.0.1,<5.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0",
  "ruff>=0.12.0,<1.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "tests", "pyproject.toml", "README.md", ".env.example"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
